codegen: split gengo.go file
